core: https and quic carriers on Windows (static OpenSSL 3.5)
The Windows core links OpenSSL 3.5.7, zlib, zstd and ngtcp2 statically
(Dockerfile.build-win, Debian 13) instead of libgcrypt, so tincd.exe
serves and dials both TLS carriers with the Linux build's wire image.

Fixes found on the way:
- https dial socket stayed blocking on Windows (no O_NONBLOCK): FIONBIO
- a dial to an IP sent SNI and Host 'localhost'; now no SNI and
  Host <ip>[:port], as curl does
- %zd/%zu formats, memmem/strcasestr fallbacks, NGTCP2_STATICLIB
- Windows GUI wrote HttpsFront/HttpsFrontPort, which no core reads;
  it now writes HttpsPort (0 = off)

Proof: testing/transports/windows-wine-test.sh 14/14 under Wine;
Linux regressions 24/24.

+++ core/tincd/src/httpc.c
@@ -384,7 +384,7 @@ bool httpc_request(const httpc_request_t *req, http_response_t *res, char *err, 
 		goto fail;
 	}
 
-	n += snprintf(head + n, sizeof(head) - (size_t) n, "Content-Length: %zu\r\n\r\n", req->body_len);
+	n += snprintf(head + n, sizeof(head) - (size_t) n, "Content-Length: %lu\r\n\r\n", (unsigned long)req->body_len);
 
 	if(SSL_write(ssl, head, n) != n) {
 		ssl_error(err, errlen, "cannot send the request", ssl);
